refactor(Anizium): Modularize presence script and add logo setting
This commit refactors the Anizium presence script by breaking down the monolithic `presence.ts` file into smaller, single-responsibility modules. This improves code organization, readability, and maintainability.

New modules introduced:
- `route-handlers.ts`: Manages logic for different URL paths.
- `settings-manager.ts`: Handles fetching and storing user settings.
- `poster-manager.ts`: Manages fetching anime poster images.
- `utils.ts`: Contains shared helper functions.
- `types.ts`: Defines shared TypeScript interfaces.

Additionally, this commit adds a new "Icon" setting, allowing users to choose between an animated, still, or no-background logo for the presence.
